requirements/base.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-decouple==3.8
sqlparse==0.5.2
